Made small changes to rearrange code.
