+++ app/src/lib/formSchemas.ts
@@ -1,5 +1,13 @@
 import { z } from "zod";
 
+export const US_STATES = [
+  "AL", "AK", "AZ", "AR", "CA", "CO", "CT", "DE", "FL", "GA",
+  "HI", "ID", "IL", "IN", "IA", "KS", "KY", "LA", "ME", "MD",
+  "MA", "MI", "MN", "MS", "MO", "MT", "NE", "NV", "NH", "NJ",
+  "NM", "NY", "NC", "ND", "OH", "OK", "OR", "PA", "RI", "SC",
+  "SD", "TN", "TX", "UT", "VT", "VA", "WA", "WV", "WI", "WY"
+];
+
 export const consentSchema = z.object({
   consentGiven: z.boolean().refine((val) => val === true, {
     message: "You must agree to share your address to continue",
